Consolidate mock OEIS tests
