+++ flaskr/nscope/test/test_wrong_response_type.py
@@ -1,401 +1,23 @@
-import os
-import unittest
-import requests
-import requests.exceptions
-import sys
-import time
-import mock_oeis
-from urllib.parse import urlunparse
-from subprocess import Popen
-from flaskr import create_app, db
-import flaskr.nscope.views as views
+import flaskr.nscope.test.abstract_mock_oeis_test as abstract_mock_oeis_test
 
-
-# for starting and stopping the server, hat tip StackOverflow user moe asal...
-#   https://stackoverflow.com/q/63422975
-# ... who provided code under the MIT license, and ideas under CC BY-SA 4.0
-#   https://meta.stackexchange.com/q/271080
-
-
-class TestUnavailableValues(unittest.TestCase):
-  mock_scheme = 'http'
-  mock_host = 'localhost'
-  mock_port = '5001'
-  mock_hostport = f'{mock_host}:{mock_port}'
-
+class TestUnavailableValues(abstract_mock_oeis_test.AbstractMockOEISTest):
+  values_available = False
   endpoint = 'http://localhost:5000/api/get_oeis_values/A153080/12'
-  expected_response_text = 'Error: 503 Server Error: SERVICE UNAVAILABLE for url: http://localhost:5001/A153080/b153080.txt'
-
-  def mock_oeis_url(self, path):
-    return urlunparse([
-      self.mock_scheme,
-      self.mock_hostport,
-      path,
-      '', # path parameters
-      '', # query
-      ''  # fragment
-    ])
-
-  def waitUntilReady(self):
-    # wait for the server to start up
-    interval = 0.25
-    retry = True
-    while retry:
-      if interval > 2:
-        raise RuntimeError('Mock OEIS server took too long to start')
-      time.sleep(interval)
-      if self.verbose:
-        print(f'  Waited {interval}s; attempting connection to mock OEIS server')
-      try:
-        response = requests.get(self.mock_oeis_url('/ready'))
-        retry = not response.status_code == 200
-      except requests.exceptions.ConnectionError:
-        retry = True
-      interval = 2*interval
-
-  def setUp(self):
-    # check whether unittest is running in verbose mode
-    # hat tip StackOverflow users Dimitris Fasarakis Hilliard and EquipDev...
-    #   https://stackoverflow.com/a/43002355
-    #   https://stackoverflow.com/questions/43001768/how-can-a-test-in-python-unittest-get-access-to-the-verbosity-level#comment73163492_43002355
-    # ... who provided this code under the MIT license
-    #   https://meta.stackexchange.com/q/271080
-    self.verbose = ('-v' in sys.argv) or ('--verbose' in sys.argv)
-    if self.verbose:
-      print()
-
-    # create app
-    self.app = create_app('testing', oeis_scheme='http', oeis_hostport=f'localhost:{self.mock_port}')
-    self.ctx = self.app.app_context()
-    with self.ctx:
-      db.drop_all()
-      db.create_all()
-
-    # start serving mock OEIS
-    self.oeis_server = Popen(
-      ['flask', '--app', 'mock_oeis', 'run', '--port', self.mock_port],
-      env={'VALUES_AVAILABLE': '0', **os.environ}
-    )
-
-  def tearDown(self):
-    # wait for background work to finish
-    if self.verbose:
-      print("  Waiting for background work")
-    views.executor.shutdown()
-    if self.verbose:
-      print("  Background work done")
-
-    # clear database
-    db.session.remove()
-    with self.ctx:
-      db.drop_all()
-
-    # send termination signal to server thread
-    self.oeis_server.terminate()
-
-    # wait for server subprocess to terminate
-    self.oeis_server.wait()
-
-  def test_endpoint(self):
-    # using test client is recommended in Flask testing how-to
-    #   https://flask.palletsprojects.com/en/2.3.x/testing/
-    # "The test client makes requests to the application without running a live
-    # server." the `with` block runs teardown
-    #   https://github.com/pallets/flask/issues/2949
-    self.waitUntilReady()
-    with self.app.test_client() as client:
-      if self.verbose:
-        print("  Testing response")
-      response = client.get(self.endpoint)
-      self.assertEqual(response.status_code, 200)
-      self.assertEqual(response.text, self.expected_response_text)
+  expected_response = 'Error: 503 Server Error: SERVICE UNAVAILABLE for url: http://localhost:5001/A153080/b153080.txt'
 
 
-class TestUnavailableValuesInNameAndValues(unittest.TestCase):
-  mock_scheme = 'http'
-  mock_host = 'localhost'
-  mock_port = '5001'
-  mock_hostport = f'{mock_host}:{mock_port}'
-
+class TestUnavailableValuesInNameAndValues(abstract_mock_oeis_test.AbstractMockOEISTest):
+  values_available = False
   endpoint = 'http://localhost:5000/api/get_oeis_name_and_values/A153080'
-  expected_response_text = 'Error: 503 Server Error: SERVICE UNAVAILABLE for url: http://localhost:5001/A153080/b153080.txt'
-
-  def mock_oeis_url(self, path):
-    return urlunparse([
-      self.mock_scheme,
-      self.mock_hostport,
-      path,
-      '', # path parameters
-      '', # query
-      ''  # fragment
-    ])
-
-  def waitUntilReady(self):
-    # wait for the server to start up
-    interval = 0.25
-    retry = True
-    while retry:
-      if interval > 2:
-        raise RuntimeError('Mock OEIS server took too long to start')
-      time.sleep(interval)
-      if self.verbose:
-        print(f'  Waited {interval}s; attempting connection to mock OEIS server')
-      try:
-        response = requests.get(self.mock_oeis_url('/ready'))
-        retry = not response.status_code == 200
-      except requests.exceptions.ConnectionError:
-        retry = True
-      interval = 2*interval
-
-  def setUp(self):
-    # check whether unittest is running in verbose mode
-    # hat tip StackOverflow users Dimitris Fasarakis Hilliard and EquipDev...
-    #   https://stackoverflow.com/a/43002355
-    #   https://stackoverflow.com/questions/43001768/how-can-a-test-in-python-unittest-get-access-to-the-verbosity-level#comment73163492_43002355
-    # ... who provided this code under the MIT license
-    #   https://meta.stackexchange.com/q/271080
-    self.verbose = ('-v' in sys.argv) or ('--verbose' in sys.argv)
-    if self.verbose:
-      print()
-
-    # create app
-    self.app = create_app(
-      'testing',
-      oeis_scheme='http',
-      oeis_hostport=f'localhost:{self.mock_port}'
-    )
-    self.ctx = self.app.app_context()
-    with self.ctx:
-      db.drop_all()
-      db.create_all()
-
-    # start serving mock OEIS
-    self.oeis_server = Popen(
-      ['flask', '--app', 'mock_oeis', 'run', '--port', self.mock_port],
-      env={'VALUES_AVAILABLE': '0', **os.environ}
-    )
-
-  def tearDown(self):
-    # wait for background work to finish
-    if self.verbose:
-      print("  Waiting for background work")
-    views.executor.shutdown()
-    if self.verbose:
-      print("  Background work done")
-
-    # clear database
-    db.session.remove()
-    with self.ctx:
-      db.drop_all()
-
-    # send termination signal to server thread
-    self.oeis_server.terminate()
-
-    # wait for server subprocess to terminate
-    self.oeis_server.wait()
-
-  def test_endpoint(self):
-    # using test client is recommended in Flask testing how-to
-    #   https://flask.palletsprojects.com/en/2.3.x/testing/
-    # "The test client makes requests to the application without running a live
-    # server." the `with` block runs teardown
-    #   https://github.com/pallets/flask/issues/2949
-    self.waitUntilReady()
-    with self.app.test_client() as client:
-      if self.verbose:
-        print("  Testing response")
-      response = client.get(self.endpoint)
-      self.assertEqual(response.status_code, 200)
-      self.assertEqual(response.text, self.expected_response_text)
+  expected_response = 'Error: 503 Server Error: SERVICE UNAVAILABLE for url: http://localhost:5001/A153080/b153080.txt'
 
 
-class TestUnavailableSearchInNameAndValues(unittest.TestCase):
-  mock_scheme = 'http'
-  mock_host = 'localhost'
-  mock_port = '5001'
-  mock_hostport = f'{mock_host}:{mock_port}'
-
+class TestUnavailableSearchInNameAndValues(abstract_mock_oeis_test.AbstractMockOEISTest):
+  search_available = False
   endpoint = 'http://localhost:5000/api/get_oeis_name_and_values/A153080'
-  expected_response_text = 'Error: 503 Server Error: SERVICE UNAVAILABLE for url: http://localhost:5001/search?id=A153080&fmt=json'
+  expected_response = 'Error: 503 Server Error: SERVICE UNAVAILABLE for url: http://localhost:5001/search?id=A153080&fmt=json'
 
-  def mock_oeis_url(self, path):
-    return urlunparse([
-      self.mock_scheme,
-      self.mock_hostport,
-      path,
-      '', # path parameters
-      '', # query
-      ''  # fragment
-    ])
-
-  def waitUntilReady(self):
-    # wait for the server to start up
-    interval = 0.25
-    retry = True
-    while retry:
-      if interval > 2:
-        raise RuntimeError('Mock OEIS server took too long to start')
-      time.sleep(interval)
-      if self.verbose:
-        print(f'  Waited {interval}s; attempting connection to mock OEIS server')
-      try:
-        response = requests.get(self.mock_oeis_url('/ready'))
-        retry = not response.status_code == 200
-      except requests.exceptions.ConnectionError:
-        retry = True
-      interval = 2*interval
-
-  def setUp(self):
-    # check whether unittest is running in verbose mode
-    # hat tip StackOverflow users Dimitris Fasarakis Hilliard and EquipDev...
-    #   https://stackoverflow.com/a/43002355
-    #   https://stackoverflow.com/questions/43001768/how-can-a-test-in-python-unittest-get-access-to-the-verbosity-level#comment73163492_43002355
-    # ... who provided this code under the MIT license
-    #   https://meta.stackexchange.com/q/271080
-    self.verbose = ('-v' in sys.argv) or ('--verbose' in sys.argv)
-    if self.verbose:
-      print()
-
-    # create app
-    self.app = create_app('testing', oeis_scheme='http', oeis_hostport=f'localhost:{self.mock_port}')
-    self.ctx = self.app.app_context()
-    with self.ctx:
-      db.drop_all()
-      db.create_all()
-
-    # start serving mock OEIS
-    self.oeis_server = Popen(
-      ['flask', '--app', 'mock_oeis', 'run', '--port', self.mock_port],
-      env={'SEARCH_AVAILABLE': '0', **os.environ}
-    )
-
-  def tearDown(self):
-    # wait for background work to finish
-    if self.verbose:
-      print("  Waiting for background work")
-    views.executor.shutdown()
-    if self.verbose:
-      print("  Background work done")
-
-    # clear database
-    db.session.remove()
-    with self.ctx:
-      db.drop_all()
-
-    # send termination signal to server thread
-    self.oeis_server.terminate()
-
-    # wait for server subprocess to terminate
-    self.oeis_server.wait()
-
-  def test_endpoint(self):
-    # using test client is recommended in Flask testing how-to
-    #   https://flask.palletsprojects.com/en/2.3.x/testing/
-    # "The test client makes requests to the application without running a live
-    # server." the `with` block runs teardown
-    #   https://github.com/pallets/flask/issues/2949
-    self.waitUntilReady()
-    with self.app.test_client() as client:
-      if self.verbose:
-        print("  Testing response")
-      response = client.get(self.endpoint)
-      self.assertEqual(response.status_code, 200)
-      self.assertEqual(response.text, self.expected_response_text)
-
-
-class TestUnavailableSearch(unittest.TestCase):
-  mock_scheme = 'http'
-  mock_host = 'localhost'
-  mock_port = '5001'
-  mock_hostport = f'{mock_host}:{mock_port}'
-
+class TestUnavailableSearch(abstract_mock_oeis_test.AbstractMockOEISTest):
+  search_available = False
   endpoint = 'http://localhost:5000/api/get_oeis_metadata/A153080'
-  expected_response_text = 'Error: 503 Server Error: SERVICE UNAVAILABLE for url: http://localhost:5001/search?q=A153080&fmt=json'
-
-  def mock_oeis_url(self, path):
-    return urlunparse([
-      self.mock_scheme,
-      self.mock_hostport,
-      path,
-      '', # path parameters
-      '', # query
-      ''  # fragment
-    ])
-
-  def waitUntilReady(self):
-    # wait for the server to start up
-    interval = 0.25
-    retry = True
-    while retry:
-      if interval > 2:
-        raise RuntimeError('Mock OEIS server took too long to start')
-      time.sleep(interval)
-      if self.verbose:
-        print(f'  Waited {interval}s; attempting connection to mock OEIS server')
-      try:
-        response = requests.get(self.mock_oeis_url('/ready'))
-        retry = not response.status_code == 200
-      except requests.exceptions.ConnectionError:
-        retry = True
-      interval = 2*interval
-
-  def setUp(self):
-    # check whether unittest is running in verbose mode
-    # hat tip StackOverflow users Dimitris Fasarakis Hilliard and EquipDev...
-    #   https://stackoverflow.com/a/43002355
-    #   https://stackoverflow.com/questions/43001768/how-can-a-test-in-python-unittest-get-access-to-the-verbosity-level#comment73163492_43002355
-    # ... who provided this code under the MIT license
-    #   https://meta.stackexchange.com/q/271080
-    self.verbose = ('-v' in sys.argv) or ('--verbose' in sys.argv)
-    if self.verbose:
-      print()
-
-    # create app
-    self.app = create_app('testing', oeis_scheme='http', oeis_hostport=f'localhost:{self.mock_port}')
-    self.ctx = self.app.app_context()
-    with self.ctx:
-      db.drop_all()
-      db.create_all()
-
-    # start serving mock OEIS
-    self.oeis_server = Popen(
-      ['flask', '--app', 'mock_oeis', 'run', '--port', self.mock_port],
-      env={'SEARCH_AVAILABLE': '0', **os.environ}
-    )
-
-  def tearDown(self):
-    # wait for background work to finish
-    if self.verbose:
-      print("  Waiting for background work")
-    views.executor.shutdown()
-    if self.verbose:
-      print("  Background work done")
-
-    # clear database
-    db.session.remove()
-    with self.ctx:
-      db.drop_all()
-
-    # send termination signal to server thread
-    self.oeis_server.terminate()
-
-    # wait for server subprocess to terminate
-    self.oeis_server.wait()
-
-  def test_endpoint(self):
-    # using test client is recommended in Flask testing how-to
-    #   https://flask.palletsprojects.com/en/2.3.x/testing/
-    # "The test client makes requests to the application without running a live
-    # server." the `with` block runs teardown
-    #   https://github.com/pallets/flask/issues/2949
-    self.waitUntilReady()
-    with self.app.test_client() as client:
-      if self.verbose:
-        print("  Testing response")
-      response = client.get(self.endpoint)
-      self.assertEqual(response.status_code, 200)
-      self.assertEqual(response.text, self.expected_response_text)
-
-
-if __name__ == "__main__":
-  unittest.main()
+  expected_response = 'Error: 503 Server Error: SERVICE UNAVAILABLE for url: http://localhost:5001/search?q=A153080&fmt=json'
